refactor: extract tooltip core CSS to styles folder
