+++ packages/tooltip/src/styles/vaadin-tooltip-overlay-base-styles.d.ts
