+++ packages/tooltip/src/styles/vaadin-tooltip-overlay-base-styles.js
@@ -1,0 +1,8 @@
+/**
+ * @license
+ * Copyright (c) 2022 - 2025 Vaadin Ltd.
+ * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
+ */
+import { css } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
+
+export const tooltipOverlayStyles = css``;
