+++ packages/tooltip/src/styles/vaadin-tooltip-overlay-core-styles.js
