+++ packages/tooltip/src/vaadin-tooltip-overlay.js
@@ -11,7 +11,7 @@ import { overlayStyles } from '@vaadin/overlay/src/vaadin-overlay-styles.js';
 import { PopoverOverlayMixin } from '@vaadin/popover/src/vaadin-popover-overlay-mixin.js';
 import { CSSInjectionMixin } from '@vaadin/vaadin-themable-mixin/css-injection-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
-import { tooltipOverlayStyles } from './vaadin-tooltip-overlay-styles.js';
+import { tooltipOverlayStyles } from './styles/vaadin-tooltip-overlay-core-styles.js';
 
 /**
  * An element used internally by `<vaadin-tooltip>`. Not intended to be used separately.
